+++ packages/card/src/components/CheckCard/demos/description.tsx
@@ -1,4 +1,6 @@
-/** Uuid: 15886807 title: 自定义描述 */
+/**
+ * title: 自定义描述
+ */
 
 import React from 'react';
 import { CheckCard } from '@ant-design/pro-card';