+++ packages/form/src/components/SchemaForm/demos/dynamic-rerender.tsx
@@ -1,0 +1,121 @@
+import React from 'react';
+import type { ProFormColumnsType } from '@ant-design/pro-form';
+import { BetaSchemaForm } from '@ant-design/pro-form';
+import { Input } from 'antd';
+
+const valueEnum = {
+  all: { text: '全部', status: 'Default' },
+  open: {
+    text: '未解决',
+    status: 'Error',
+  },
+  closed: {
+    text: '已解决',
+    status: 'Success',
+    disabled: true,
+  },
+  processing: {
+    text: '解决中',
+    status: 'Processing',
+  },
+};
+
+type DataItem = {
+  name: string;
+  state: string;
+  title: string;
+};
+
+const columns: ProFormColumnsType<DataItem>[] = [
+  {
+    title: '标题',
+    dataIndex: 'title',
+    initialValue: '必填',
+    formItemProps: {
+      rules: [
+        {
+          required: true,
+          message: '此项为必填项',
+        },
+      ],
+    },
+    width: 'm',
+  },
+  {
+    title: '状态',
+    dataIndex: 'state',
+    valueType: 'select',
+    valueEnum,
+    width: 'm',
+    tooltip: '当title为disabled时状态无法选择',
+    fieldProps: (form) => {
+      if (form.getFieldValue('title') === 'disabled') {
+        return {
+          disabled: true,
+          placeholder: 'disabled',
+        };
+      } else {
+        return {
+          placeholder: 'normal',
+        };
+      }
+    },
+  },
+  {
+    title: '标签',
+    dataIndex: 'labels',
+    width: 'm',
+    tooltip: '当title为必填时此项将为必填',
+    formItemProps(form) {
+      if (form.getFieldValue('title') === '必填') {
+        return {
+          rules: [
+            {
+              required: true,
+            },
+          ],
+        };
+      } else {
+        return {};
+      }
+    },
+  },
+  {
+    title: 'title为hidden时隐藏',
+    dataIndex: 'hidden',
+    valueType: 'date',
+    renderFormItem: (_, __, form) => {
+      if (form.getFieldValue('title') === 'hidden') {
+        return false;
+      } else {
+        return <Input />;
+      }
+    },
+  },
+  {
+    title: '创建时间',
+    key: 'showTime',
+    dataIndex: 'createName',
+    valueType: 'date',
+  },
+];
+
+export default () => {
+  return (
+    <>
+      <BetaSchemaForm<DataItem>
+        shouldUpdate={(newValues, oldValues) => {
+          if (newValues.title !== oldValues?.title) {
+            return true;
+          }
+          return false;
+        }}
+        layoutType="Form"
+        onFinish={async (values) => {
+          console.log(values);
+        }}
+        columns={columns}
+      />
+    </>
+  );
+};