+++ packages/form/src/layouts/StepsForm/index.tsx
@@ -12,7 +12,7 @@ import type { StepFormProps } from './StepForm';
 import StepForm from './StepForm';
 import './index.less';
 import type { ProFormProps } from '../ProForm';
-import type { SubmitterProps } from '../../components/Submitter';
+import type { SubmitterProps } from '../../components';
 
 type StepsFormProps<T = Record<string, any>> = {
   /**
@@ -189,6 +189,7 @@ function StepsForm<T = Record<string, any>>(
     from.current?.submit();
   };
 
+  /** 上一页功能 */
   const prePage = useRefFunction(() => {
     if (step < 1) return;
     setStep(step - 1);
@@ -371,4 +372,4 @@ function StepsFormWarp<T = Record<string, any>>(
 StepsFormWarp.StepForm = StepForm;
 StepsFormWarp.useForm = Form.useForm;
 
-export default StepsFormWarp;
+export { StepsFormWarp as StepsForm };