+++ packages/list/src/demos/card-list.tsx
@@ -61,8 +61,19 @@ export default () => {
           defaultPageSize: 8,
           showSizeChanger: false,
         }}
+        showActions="hover"
         rowSelection={{}}
         grid={{ gutter: 16, column: 2 }}
+        onItem={(record: any) => {
+          return {
+            onMouseEnter: () => {
+              console.log(record);
+            },
+            onClick: () => {
+              console.log(record);
+            },
+          };
+        }}
         metas={{
           title: {},
           subTitle: {},