+++ packages/utils/src/hooks/useRefFunction/index.ts
@@ -1,11 +1,11 @@
-﻿import { useRef } from 'react';
+﻿import { useCallback, useRef } from 'react';
 
 const useRefFunction = <T extends (...args: any) => any>(reFunction: T) => {
   const ref = useRef<any>(null);
   ref.current = reFunction;
-  return (...rest: Parameters<T>): ReturnType<T> => {
+  return useCallback((...rest: Parameters<T>): ReturnType<T> => {
     return ref.current?.(...(rest as any));
-  };
+  }, []);
 };
 
 export { useRefFunction };