+++ packages/provider/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.5.7...@ant-design/pro-provider@1.5.8) (2022-02-14)
+
+### Bug Fixes
+
+- **field:** setting debounceTime results in repeated requests ([#4609](https://github.com/ant-design/pro-components/issues/4609)) ([51d2441](https://github.com/ant-design/pro-components/commit/51d2441b080c7060fe083803e8a89c4db3ec0a67))
+
+## [1.5.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.5.6...@ant-design/pro-provider@1.5.7) (2022-02-10)
+
+**Note:** Version bump only for package @ant-design/pro-provider
+
+## [1.5.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.5.5...@ant-design/pro-provider@1.5.6) (2022-01-21)
+
+### Bug Fixes
+
+- **form:** 🐛 readonly form will not render the latest value ([#4494](https://github.com/ant-design/pro-components/issues/4494)) ([2c169a8](https://github.com/ant-design/pro-components/commit/2c169a846b72351f9988c22fd5c2ef2b1684b577))
+
 ## [1.5.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-provider@1.5.4...@ant-design/pro-provider@1.5.5) (2022-01-05)
 
 **Note:** Version bump only for package @ant-design/pro-provider